Refs #21312 - merging i18n files
