+++ app/assets/javascripts/application.js
@@ -89,7 +89,7 @@ function onContentLoad(){
 
   password_caps_lock_hint();
 
-  tfm.i18n.ready.then(function() {
+  tfm.i18n.intl.ready.then(function() {
     var tz = jstz.determine();
     $.cookie('timezone', tz.name(), { path: '/', secure: location.protocol === 'https:' });
   });
